Fix the merge request to email bridge
This got broken when the Branch model got added to handle multiple
development branches. This also fixes an issue where merge requests got
bridged to email and then back to merge requests.

Additionally, it adds some basic missing features:

* Add in-reply-to headers

* Add authorship information to mail bodies.

Finally, it yanks the non-task code out of the task module. This isn't
an atomic commit and I am ashamed.
