+++ patchlab/tests/__init__.py
@@ -41,6 +41,8 @@ X-Patchlab-Patch-Author: Jeremy Cline <jcline@redhat.com>
 X-Patchlab-Merge-Request: https://gitlab/root/kernel/merge_requests/1
 X-Patchlab-Commit: a958a0dff5e3c433eb99bc5f18cbcfad77433b0d
 
+From: Jeremy Cline <jcline@redhat.com>
+
 This is a silly change so I can write a test.
 
 Signed-off-by: Jeremy Cline <jcline@redhat.com>
@@ -75,6 +77,8 @@ Date: Mon, 04 Nov 2019 23:00:00 -0000
 Message-ID: <1@localhost.localdomain>
 X-Patchlab-Merge-Request: https://gitlab/root/kernel/merge_requests/2
 
+From: Administrator <admin@example.com>
+
 Update the README to make me want to read it more.""",
     """Content-Type: text/plain; charset="utf-8"
 MIME-Version: 1.0
@@ -89,6 +93,9 @@ Message-ID: <2@localhost.localdomain>
 X-Patchlab-Patch-Author: Jeremy Cline <jcline@redhat.com>
 X-Patchlab-Merge-Request: https://gitlab/root/kernel/merge_requests/2
 X-Patchlab-Commit: 5c9b066a8bc9eed0e8d7ccd392bc8f77c42532f0
+In-Reply-To: <1@localhost.localdomain>
+
+From: Jeremy Cline <jcline@redhat.com>
 
 This is a silly change so I can write a test.
 
@@ -123,6 +130,9 @@ Message-ID: <4@localhost.localdomain>
 X-Patchlab-Patch-Author: Jeremy Cline <jcline@redhat.com>
 X-Patchlab-Merge-Request: https://gitlab/root/kernel/merge_requests/2
 X-Patchlab-Commit: c321c86ee75491f4bc0b0b0e368f71eff88fa91c
+In-Reply-To: <1@localhost.localdomain>
+
+From: Jeremy Cline <jcline@redhat.com>
 
 Make the README more readable.
 
@@ -186,6 +196,8 @@ Date: Mon, 04 Nov 2019 23:00:00 -0000
 Message-ID: <4@localhost.localdomain>
 X-Patchlab-Merge-Request: https://gitlab/root/kernel/merge_requests/2
 
+From: Administrator <admin@example.com>
+
 Update the README to make me want to read it more.
 
 Note:
@@ -195,12 +207,12 @@ The patch series is too large to sent by email.
 Reviewing locally, set up your repository to fetch from the GitLab remote:
 
   $ git remote add gitlab https://gitlab/root/kernel.git
-  $ git config remote.gitlab.fetch '+refs/merge-requests/*:refs/remotes/origin/merge-requests/*'
+  $ git config remote.gitlab.fetch '+refs/merge-requests/*/head:refs/remotes/gitlab/merge-requests/*'
   $ git fetch gitlab
 
 Finally, check out the merge request:
 
-  $ git checkout merge-requests/2
+  $ git checkout gitlab/merge-requests/2
 
 It is also possible to review the merge request on GitLab at:
     https://gitlab/root/kernel/merge_requests/2
