+++ patchlab/tests/fixtures/VCR/patchlab.tests.test_gitlab2email.PrepareEmailsTests.test_no_patch_prefix
