+++ patchlab/tests/fixtures/VCR/patchlab.tests.test_gitlab2email.RecordBridgingTests.test_duplicate_patches
