+++ patchlab/tests/fixtures/VCR/patchlab.tests.test_gitlab2email.RecordBridgingTests.test_single_patch_series
